requirements.txt
redis>=3.2.1
Flask>=1.0.2
Flask-JWT-Extended>=3.18.2
pytest>=4.5.0
pyyaml>=5.1